merge Dev to master

+++ src/main/java/com/shop/pbl6_shop_fashion/dto/ProductDetailMobileDto.java
@@ -3,6 +3,7 @@ package com.shop.pbl6_shop_fashion.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -17,8 +18,8 @@ public class ProductDetailMobileDto {
     private  String ProductName;
     private Long price;
     private Long price_promote;
-    private Long quantity;
-    private Long quantity_sold;
+    private BigDecimal quantity;
+    private BigDecimal quantity_sold;
     private String decription;
     private List<String> productUrls;
     private List<Date> commentCreatedAts;
